requirements.txt
flask
six
python-ldap
PyJWT
flask-login
flask-restplus
bravado
Flask-Cors
apispec
webargs
flask-marshmallow
flatdict
bidict
pymongo
Flask-PyMongo
watchdog
socketIO-client
flask-socketio
gevent
gevent-websocket
PyYAML